requirements.txt
streamlit
groq
torch
torchvision
transformers
pillow
requests
python-dotenv